run doris cluster

+++ doris_ops.sh
@@ -5,8 +5,8 @@ test  ${basedir} == ${PWD}
 dorisLocalRoot=$(cd ${basedir}/../doris_all;pwd)
 dorisDockerRoot=/root/doris
 
-doris_fe_num=$(perl -ne 'print if /^\s*\d+(\.\d+){3}\s+doris_fe\d+\s*$/' ${PWD}/hosts |wc -l);
-doris_be_num=$(perl -ne 'print if /^\s*\d+(\.\d+){3}\s+doris_be\d+\s*$/' ${PWD}/hosts |wc -l);
+doris_fe_list=$(perl -lne 'print $1 if /^\s*\d+(?:\.\d+){3}\s+(doris_fe_(:?follower|observer)\d+)\s*$/' ${PWD}/hosts )
+doris_be_list=$(perl -lne 'print $1 if /^\s*\d+(?:\.\d+){3}\s+(doris_be\d+)\s*$/' ${PWD}/hosts )
 
 dockerFlags="-tid --rm -u root --privileged --net static_net0 -v ${PWD}/hosts:/etc/hosts -v ${dorisLocalRoot}:${dorisDockerRoot}"
 
@@ -20,50 +20,111 @@ stop_node(){
 
 ## doris-fe
 
-bootstrap_doris_fe(){
-  local node=${1:?"undefined 'doris_fe'"};shift
-  stop_node ${node}
-  rm -fr ${basedir:?"undefined"}/${node}_data/*
-  rm -fr ${basedir:?"undefined"}/${node}_logs/*
-}
 
-bootstrap_all_doris_fe(){
-  for node in $(eval "echo doris_fe{0..$((${doris_fe_num}-1))}") ;do
-    bootstrap_doris_fe ${node}
-  done
+stop_doris_fe_args(){
+  local node=${1:?"undefined 'doris_fe'"};shift
+  local finalize=${1:-"false"}
+  stop_node ${node}
+  if [ "x${finalize}x" != 'xfalsex' ];then
+    if isContainerRunning doris_fe_follower;then
+      local ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
+      local editLogPort=$(perl -lne 'print $1 if /^\s*edit_log_port\s*=\s*(\b\d+\b)/' ${PWD}/doris_fe_conf/fe.conf)
+      editLogPort=${editLogPort:-9010}
+
+      # parse Role of doris-fe naming convention: doris_fe_follower0, doris_fe_observer2 and etc.
+      local role=$(echo $node | perl -lne 'print qq/\U$1\E/ if /doris_fe_(follower|observer)\d+/')
+      if [ -z "${role}" ];then
+        echo $(red_print "Role of doris-fe must be FOLLOWER|OBSERVER") >&2;
+        exit 1;
+      fi
+      ${basedir}/mysql.sh "ALTER SYSTEM DROP ${role} '${ip}:${editLogPort}';"
+    fi
+
+    [ -d "${PWD}/${node}_data" ] && rm -fr ${PWD}/${node}_data/*
+    mkdir -p ${PWD}/${node}_data
+  fi
 }
 
 stop_doris_fe(){
-  local node=${1:?"undefined 'doris_fe'"};shift
-  stop_node ${node}
+  stop_doris_fe_args ${1:?"missing 'node'"} "false"
+}
+
+destroy_doris_fe(){
+  stop_doris_fe_args ${1:?"missing 'node'"} "true"
 }
 
 stop_all_doris_fe(){
-  for node in $(eval "echo doris_fe{0..$((${doris_fe_num}-1))}") ;do
+  for node in ${doris_fe_list} ;do
     stop_doris_fe ${node}
   done
 }
 
-start_doris_fe(){
+start_doris_fe_args(){
   local node=${1:?"undefined 'doris_fe'"};shift
-	ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
-  flags="
+  local bootstrap=${1:-"false"};shift;
+  local ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
+  local flags="
   -v ${PWD}/${node}_data:${dorisDockerRoot}/fe/doris-meta
   -v ${PWD}/${node}_logs:${dorisDockerRoot}/fe/log
+  -v ${PWD}/${node}_run:${dorisDockerRoot}/fe/run
   -v ${PWD}/doris_fe_conf:${dorisDockerRoot}/fe/conf
+  -e PID_DIR=${dorisDockerRoot}/fe/run
   --name $node
   --hostname $node
   --ip $ip
   "
-  rm -fr ${PWD}/${node}_logs/*
+
+  [ -d "${PWD}/${node}_run" ] && rm -fr ${PWD}/${node}_run/*
+  mkdir -p ${PWD}/${node}_run
+  [ -d "${PWD}/${node}_logs" ] && rm -fr ${PWD}/${node}_logs/*
   mkdir -p ${PWD}/${node}_logs
-  docker run ${dockerFlags} ${flags} \
-    apachedoris/doris-dev:build-env-1.2 \
-    ${dorisDockerRoot}/fe/bin/start_fe.sh
+
+  # parse Role of doris-fe naming convention: doris_fe_follower0, doris_fe_observer2 and etc.
+  local role=$(echo $node | perl -lne 'print qq/\U$1\E/ if /doris_fe_(follower|observer)\d+/')
+  if [ -z "${role}" ];then
+    echo $(red_print "Role of doris-fe must be FOLLOWER|OBSERVER") >&2;
+    exit 1;
+  fi
+
+  # bootstrap-mode: cleanup datadir of doris-fe 
+  [ "x${bootstrap}x" != "xfalse" -a -d "${PWD}/${node}_data" ] && rm -fr ${PWD}/${node}_data/*
+  mkdir -p ${PWD}/${node}_data
+
+  # bootstrap-mode: doris-fe instances other than the first one should register itself to the first one.
+  # ALTER SYSTEM ADD FOLLOWER|OBSERVER 'self_ip:self_port'
+  # fe/bin/start_fe.sh --helper 'masterFs_ip:masterFe_port'
+  local helpOption=""
+  if [ "x${bootstrap}x" != "xfalsex" ] && isContainerRunning doris_fe_follower;then
+    echo "Another doris_fe instances have already started, so get editlog_port via mysql"
+    masterFe=$(${basedir}/mysql.sh "SHOW PROC '/frontends';" |perl -aF'\s+' -lne  'print "$F[1]:$F[3]" if /doris_fe/ && $F[8] eq qq/true/'|head -1)
+    editLogPort=$(perl -lne 'print $1 if /^\s*edit_log_port\s*=\s*(\b\d+\b)/' ${PWD}/doris_fe_conf/fe.conf)
+    editLogPort=${editLogPort:-9010}
+    ${basedir}/mysql.sh "ALTER SYSTEM ADD ${role} '${ip}:${editLogPort}'"
+    helpOption="--helper ${masterFe}"
+  else
+    echo "It is the first doris_fe instance or non-bootstrap startup: ${node}"
+  fi
+
+  # run docker
+  docker run ${dockerFlags} ${flags} apachedoris/doris-dev:build-env-1.2 ${dorisDockerRoot}/fe/bin/start_fe.sh ${helpOption}
+}
+
+bootstrap_doris_fe(){
+  start_doris_fe_args ${1:?"undefined 'node'"} "true"
+}
+
+bootstrap_all_doris_fe(){
+  for node in ${doris_fe_list} ;do
+    bootstrap_doris_fe ${node}
+  done
+}
+
+start_doris_fe(){
+  start_doris_fe_args ${1:?"undefined 'node'"} "false"
 }
 
 start_all_doris_fe(){
-  for node in $(eval "echo doris_fe{0..$((${doris_fe_num}-1))}") ;do
+  for node in ${doris_fe_list} ;do
     start_doris_fe ${node}
   done
 }
@@ -75,7 +136,7 @@ restart_doris_fe(){
 }
 
 restart_all_doris_fe(){
-  for node in $(eval "echo doris_fe{0..$((${doris_fe_num}-1))}") ;do
+  for node in ${doris_fe_list} ;do
     restart_doris_fe ${node}
   done
 }
@@ -83,33 +144,49 @@ restart_all_doris_fe(){
 #################################################################
 ## doris-be 
 
-bootstrap_doris_be(){
-  local name=$1;shift
-  stop_node ${name}
-  rm -fr ${basedir:?"undefined"}/${name}_data/*
-  rm -fr ${basedir:?"undefined"}/${name}_logs/*
-}
 
-bootstrap_all_doris_be(){
-  for node in $(eval "echo doris_be{0..$((${doris_be_num}-1))}") ;do
-    bootstrap_doris_be ${node}
-  done
+stop_doris_be_args(){
+  local node=${1:?"missing 'node'"};shift
+  local finalize=${1:-"false"}
+
+  stop_node ${node}
+  if [ "x${finalize}x" != 'xfalsex' ];then
+    if isContainerRunning doris_fe_follower;then
+      local ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
+      local heartbeatServicePort=$(perl -lne 'print $1 if /^\s*heartbeat_service_port\s*=\s*(\b\d+\b)/' ${PWD}/doris_be_conf/be.conf)
+      heartbeatServicePort=${heartbeatServicePort:-9050}
+      ${basedir}/mysql.sh "ALTER SYSTEM DROP BACKEND '${ip}:${heartbeatServicePort}';"
+    fi
+
+    [ -d "${PWD}/${node}_data" ] && rm -fr ${PWD}/${node}_data/*
+    mkdir -p ${PWD}/${node}_data
+  fi
 }
 
 stop_doris_be(){
-  local name=$1;shift
-  stop_node ${name}
+  stop_doris_be_args ${1:?"missing 'node'"} "false"
 }
 
 stop_all_doris_be(){
-  for node in $(eval "echo doris_be{0..$((${doris_be_num}-1))}") ;do
-    stop_node ${node}
+  for node in ${doris_be_list} ;do
+    stop_doris_be ${node}
   done
 }
 
-start_doris_be(){
-  local name=$1;shift
-	ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
+destroy_doris_be(){
+  stop_doris_be_args ${1:?"missing 'node'"} "true"
+}
+
+destroy_all_doris_be(){
+  for node in ${doris_be_list};do
+    destroy_doris_be ${node}
+  done
+}
+
+start_doris_be_args(){
+  local node=${1:?"missing 'node'"};shift
+  local bootstrap=${1:-"false"};shift
+  ip=$(perl -aF/\\s+/ -ne "print \$F[0] if /\b$node\b/" hosts)
   flags="
   -v ${PWD}/${node}_data:${dorisDockerRoot}/be/storage
   -v ${PWD}/${node}_logs:${dorisDockerRoot}/be/log
@@ -118,15 +195,39 @@ start_doris_be(){
   --hostname $node
   --ip $ip
   "
-  rm -fr ${PWD}/${node}_logs/*
+
+  [ -d "${PWD}/${node}_logs" ] && rm -fr ${PWD}/${node}_logs/*
   mkdir -p ${PWD}/${node}_logs
+
+  if [ "x${bootstrap}x" != "false" ];then
+    [ -d "${PWD}/${node}_data" ] && rm -fr ${PWD}/${node}_data/*
+    mkdir -p ${PWD}/${node}_data
+    heartbeatServicePort=$(perl -lne 'print $1 if /^\s*heartbeat_service_port\s*=\s*(\b\d+\b)/' ${PWD}/doris_be_conf/be.conf)
+    heartbeatServicePort=${heartbeatServicePort:-9050}
+    ${basedir}/mysql.sh "ALTER SYSTEM ADD BACKEND '${ip}:${heartbeatServicePort}';"
+  fi
+
   docker run ${dockerFlags} ${flags} \
     apachedoris/doris-dev:build-env-1.2 \
     ${dorisDockerRoot}/be/bin/start_be.sh 
+  }
+
+bootstrap_doris_be(){
+  start_doris_be_args ${1:?"missing 'node'"} "true"
+}
+
+bootstrap_all_doris_be(){
+  for node in ${doris_be_list}; do
+    bootstrap_doris_be ${node}
+  done
+}
+
+start_doris_be(){
+  start_doris_be_args ${1:?"missing 'node'"} "false"
 }
 
 start_all_doris_be(){
-  for node in $(eval "echo doris_be{0..$((${doris_be_num}-1))}") ;do
+  for node in ${doris_be_list}; do
     start_doris_be ${node}
   done
 }
@@ -138,7 +239,7 @@ restart_doris_be(){
 }
 
 restart_all_doris_be(){
-  for node in $(eval "echo doris_be{0..$((${doris_be_num}-1))}") ;do
+  for node in ${doris_be_list}; do
     restart_doris_be ${node}
   done
 }
